# docs/swagger.yaml
definitions:
  entity.FullUser:
    properties:
      age:
        type: integer
      gender:
        type: string
      id:
        type: integer
      name:
        type: string
      nationality:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
  entity.User:
    properties:
      name:
        type: string
      patronymic:
        type: string
      surname:
        type: string
    type: object
info:
  contact: {}
  description: Это API для тестового задания.
  title: Test Task
  version: "1.0"
paths:
  /add:
    post:
      consumes:
      - application/json
      description: Создает новую запись пользователя в системе
      parameters:
      - description: Данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное добавление
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Добавить пользователя
      tags:
      - Users
  /delete/{id}:
    delete:
      description: Удаляет пользователя по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Успешное удаление
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Удалить пользователя
      tags:
      - Users
  /get/age/{age}:
    get:
      description: Возвращает список пользователей указанного возраста
      parameters:
      - description: Возраст (целое число)
        in: path
        name: age
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.FullUser'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить по возрасту
      tags:
      - Users
  /get/all:
    get:
      description: Возвращает полный список пользователей в системе
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.FullUser'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить всех пользователей
      tags:
      - Users
  /get/gender/{gender}:
    get:
      description: Возвращает список пользователей с указанным полом
      parameters:
      - description: Пол (male/female)
        in: path
        name: gender
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.FullUser'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить по полу
      tags:
      - Users
  /get/id/{id}:
    get:
      description: Возвращает полную информацию о пользователе по его идентификатору
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.FullUser'
        "400":
          description: Неверный формат ID
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить по ID
      tags:
      - Users
  /get/name/{name}:
    get:
      description: Возвращает список пользователей с указанным именем
      parameters:
      - description: Имя пользователя
        in: path
        name: name
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.FullUser'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить по имени
      tags:
      - Users
  /get/nationality/{nationality}:
    get:
      description: Возвращает список пользователей с указанной национальностью
      parameters:
      - description: Код национальности (2 заглавные буквы)
        in: path
        name: nationality
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.FullUser'
            type: array
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Получить по национальности
      tags:
      - Users
  /update/{id}:
    put:
      consumes:
      - application/json
      description: Обновляет информацию о пользователе по его ID
      parameters:
      - description: ID пользователя
        in: path
        name: id
        required: true
        type: integer
      - description: Новые данные пользователя
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/entity.User'
      produces:
      - application/json
      responses:
        "200":
          description: Успешное обновление
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Неверные данные
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Внутренняя ошибка сервера
          schema:
            additionalProperties: true
            type: object
      summary: Обновить пользователя
      tags:
      - Users
swagger: "2.0"